requirements.txt
streamlit
websocket-client
sounddevice
numpy
scipy